Dynamic Schema Generator
This PR attempts to defined a dynamic scheme to pass to OpenAI to enforce a valid set of blocks being returned. It has a couple of limitations because the OpenAI API will only work with 'strict' schemas, so you can't use `$ref` to refer to other properties and everything must be 100% explicit and defined. _Also_, the API is limited to 100 properties, so when thinking of blocks you can easily exceed 100 properties. As a result, a _very_ limited set of blocks are in this PoC, with minimal support for nested blocks.